+++ src/pages/support-us/sponsors.js
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 import styled from 'styled-components'
-import BannerImage from '../../components/bannerImage'
+import BannerImage from '../../components/banner/bannerImage'
 import SponsorsSubsection from '../../features/sponsors/components/sponsorSubsection/index'
 import constants from '../../constants'
 import theme from '../../theme/theme'
